ITC thread and window finished, still unsafe
What's missing: In case a variable is not being set, because there was an interfering query on the serial connection, the command to set this variable should be resend, until it is thouroughly set. In case that is, a visaerror is raised, which must should not just only be logged.