Plotting Window
So, i created a new window with Comboboxes, where the user can choose from the column names in the connected database, and pyplot will do its thing by opening a new window. If we want to handle multiple plots, matplotlib has a subplot function which might be useful. For now it is pretty basic but works. Check it out when you have some time.